Changes to user and project models. Changes to media folder
Whenever a user is saved and has no avatar, a default avatar is set. Whenever a project is saved
and has no logo, a default logo is set. All logos and avatars are saved in media/images folder. After a project
is deleted it's logo is deleted as well. Finally added a success message after import.